# Conflicts:
#	src/main/java/org/example/codesix/domain/card/repository/CardRepository.java

+++ src/main/java/org/example/codesix/domain/card/service/CardService.java
@@ -1,7 +1,14 @@
 package org.example.codesix.domain.card.service;
 
+import org.example.codesix.domain.card.dto.CardDetailsResponseDto;
+import org.example.codesix.domain.card.dto.CardHistoryResponseDto;
+import org.example.codesix.domain.card.entity.CardHistory;
+import org.example.codesix.domain.card.entity.CardMember;
+import org.example.codesix.domain.comment.dto.CommentResponseDto;
+import org.example.codesix.domain.user.entity.User;
 import org.example.codesix.domain.worklist.entity.WorkList;
 import org.example.codesix.domain.worklist.repository.WorkListRepository;
+import org.example.codesix.global.exception.ForbiddenException;
 import org.springframework.transaction.annotation.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.example.codesix.domain.card.dto.CardRequestDto;
@@ -15,6 +22,7 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDate;
+import java.util.List;
 
 
 @Service
@@ -35,21 +43,54 @@ public class CardService {
         return cards.map(CardResponseDto::toDto);
     }
 
-    public CardResponseDto findCard(Long workListId, Long id) {
-        Card card = cardRepository.findWorkAndList(id, workListId);
-        return CardResponseDto.toDto(card);
+    @Transactional(readOnly = true)
+    public CardDetailsResponseDto findCard(Long workListId, Long id) {
+        cardRepository.findWorkAndList(workListId,id);
+        Card card = cardRepository.findCardWithDetails(id, workListId);
+        List<Long> userIds = card.getCardMembers().stream()
+                .map(member -> member.getMember().getId())
+                .toList();
+
+        // 카드와 댓글 데이터를 DTO로 변환
+        List<CommentResponseDto> comments = card.getComments().stream()
+                .map(CommentResponseDto::toDto)
+                .toList();
+
+        // 댓글 데이터를 포함하여 CardResponseDto 반환
+        return CardDetailsResponseDto.toDtoWithComments(card,userIds, comments);
     }
 
     @Transactional
-    public CardResponseDto updateCard(Long workListId, Long id, CardRequestDto requestDto) {
+    public CardResponseDto updateCard(Long workListId, User user, Long id, CardRequestDto requestDto) {
         Card card = cardRepository.findWorkAndList(id, workListId);
+        validateCardOwner(card, user);
+        card.addHistory(createCardHistory(card, "카드 수정", user.getId()));
         card.update(requestDto.getTitle(), requestDto.getDescription(), requestDto.getDueDate());
         return CardResponseDto.toDto(card);
     }
 
-
-    public void deleteCard(Long workListId, Long id) {
+    public void deleteCard(Long workListId, User user, Long id) {
         Card card = cardRepository.findWorkAndList(id, workListId);
+        validateCardOwner(card, user);
         cardRepository.delete(card);
+    }
+
+    private CardHistory createCardHistory(Card card, String message, Long userId) {
+        return new CardHistory(card, message, userId);
+    }
+
+    private void validateCardOwner(Card card, User user) {
+        boolean isOwner = card.getCardMembers().stream()
+                .anyMatch(member -> member.getMember().getUser().getId().equals(user.getId()));
+        if (!isOwner) {
+            throw new ForbiddenException(ExceptionType.FORBIDDEN_ACTION);
+        }
+    }
+
+    public List<CardHistoryResponseDto> getHistoryByCardId(Long cardId) {
+        List<CardHistory> history = cardRepository.findHistoryByCardId(cardId);
+        return history.stream()
+                .map(CardHistoryResponseDto::toDto)
+                .toList();
     }
 }
